Add 3 new queries + update vacuum_stats view with `manual_vacuum_count` and `autovacuum_count` (#33)
* Add 3 new queries

- ANALYZE progress
Useful to check to know if PG is acquiring sample rows on a certain
table

- VACUUM progress
Useful to know when vacuum workers are working or if you suspect the
system beingn slow for some reason. This could also help check if there are
multiple vacuums happening at the same time

-  Cumulative I/O statistics with pg_stat_io for vacuum purposes
this can help detect if we're vacuuming too often. This might get added
to the diagnose page in a later PR.
More info: https://pganalyze.com/blog/pg-stat-io#tracking-cumulative-io-activity-by-autovacuum-and-manual-vacuums

* Add missing useful columns in vacuum_stats.sql

* Update README with analyze_progress / vacuum_progress / vacuum_io_stats queries + updated vacuum_stats command